test: try mocking random package (#17)
